+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## vNEXT
 - Update requester/voucher debits to refund on claims amounts divisible by volume. (#28)
+- Add slither suggestions. (#26)
+- Drain expired vouchers and withdraw funds. (#25)
 - Add slither github action. (#24)
 - Top up voucher. (#23)
 - Claim task part 2 - Add voucher tests. (#21)